CONCPP-17 Some potential coding bugs
Item 1 is not an issue, but could be written more accurately. Item 2 is
fixed, but it both lines are in the same unused function stub. Item 3
was addressed in other ticket - I believe it's the same issue as
CONCPP-14. Did nothing about 4th point, but I guess it's ok for now.