CONCPP-28 Various fixes int DatabaseMetaData class and in related tests.
The biggest class of changes is making MariaDB schema(database) a schema
in the connector, and not a catalog.
Then many methods and/or tests returned/expected wrong data.
Fixed also issue with resultset navigation. Certain sequences of
the cursor navigation methods could cause the cursor to be at the wrong
position. If I am not mistaken, one of sequences could be next-set of
navigation commands-absolute(0)-next