CONCPP-61 Fix of build with VS2015
It choked on const initialization with constexpr

Second possible issue was with
REASON_UNKNOWN in the ClientInfoStatus enum. It clashes with define in
one of Windows headers(Winreg.h). Added undersore prefix to the name.

Also fixed most often warning there.

Additionaly the commit contains some fixes in test faramework and in
tests.