CONCPP-63 Fix of possible memory leak with exceptions
The problem was, that if the created exception was a subclass as SQLException, it(or it's
copy) would eventually be destructed as SQLException. Thus, everything
but SQLException would leak. That happened, because as a quick solution
at the time, they all were thrown as SQLException. Caught as
SQLException&(they could not be caught as an exception of its own
class), they would be destructed as SQLException.

As the solution, exception are thrown at the moment of creation, if
possible. Otherwise connector operates spesially created
class(MariaDBExceptionThrower), that
can throw exception of the correct type