CONCPP-72 Second approach. Unblocking the connection if needed.
Still fetchSize 1, but if other command needs to be sent to the server,
remaining rows are fetched and cached in the Connector, and connection
is unblocked in this way for new operations.
Changing ResultSet type to other, than TYPE_FORWARD_ONLY, will reset
fetchSize, and thus RS streaming won't be done.
If fetchSize is set to non-zero, and RS type is not TYPE_FORWARD_ONLY,
the exception will be thrown, that such combination is not supported.