CONCPP-91 utf8mb4 is made default connection character set
If useCharacterEncoding(or useCharset, or OPT_SET_CHARSET_NAME) is not
specified by the application, utf8mb4 should be used.

Connector does not do any re-coding. But it assumes all string data
coming from the application to be encoded in this charset, and return
data to application also encoded using this charset.

Application still may specify the charset of its choice to use.