Fix of the bug in the Pool
Some properties were not initialized, and that led to wrong pool behavior.
Fixed couple of memory leaks in the pool - ScheduledFuture property was not deleted,
scheduled taks object could be leaked in the queue and connection object could be
leaked in the addConnection.
Changed ScheduledFuture to have weak_ptr on the task's property.

Seemingly, atomic_bool and atomic<bool> are different things for some older
gcc versions. Made use only one of them around one of classes.

Test of idle connection remover.