CONCPP-106 Correction for the main fix
In case of CSPS, connector prepared for the bulk query all changed to
lower case. Besides being wrong in general, that could cause errors on
platforms, where names are case sensitive.
Also made few methods(including batch execution methods) to unlock the
mutex in case of error, because some routines processing error also try
to acquire the lock, and it is not re-entrant.