CONCPP-111 Introducing prepared queries cache.
Realization has been moved from a class, that existed in the source tree, but was
not used, to a external class template. The cache is controlled by
following connection string options:
 - cachePrepStmts - enables/disables use of cache, by default is false
 - prepStmtCacheSize - sets the size of cache, by default 250
 - prepStmtCacheSqlLimit - allows to change maximum size of the query,
  that will be cached. This value consists of length of query itself +
  length of schema name + 1
Cache will be enables only of if useServerPrepStmts is selected.
Test of PS caching has been added, however it can fail if run on not
exclusive server(i.e. used by other apps in parallel)
Update of travis matrix