CONCPP-117 Second part
Introduced MARIADB_STATIC_LINK that needs to be defined if application
want to link static C/C++ lib on Windows. The other way is define
MARIADB_EXPORTS and MARRIADB_EXTERN as extern

Made MARIADB_EXPORTS definition in cmake target specific, and not
project wide.
Not sure why, but submodule looks to be not updated. Moved to v3.3.8 tag