Changes to make MinGW work (PR #11)
---------

Co-authored-by: Lawrin Novitsky <lawrin.novitsky@mariadb.com>

I've made things simpler. Hopefully I've understood correctly what
problems did mingw have.

Along the way I've chcnged priority of the USE_SYSTEM_INSTALLED_LIB. It
wasn't respected, if libmariadb subdirecory is present and there is
CMakeLIsts.txt in it. Now if it's set, connector will be looking for
libs and headers installed on the system.