[CMake] Make possible building as submodule
The CMake variable `CMAKE_SOURCE_DIR` expands to the directory of the
top level CMakeLists being invoked. When added as a git submodule and
imported into a CMake project via `add_subdirectory`,
mariadb-connector-cpp fails to configure as it attempts to look for its
files in the top level directory instead of its own directory. For
example, assuming mariadb-connector-cpp is in
`project/ext/mariadb-connector-cpp/`, when attempting to find a file X,
the search path is `project/X` instead of
`project/ext/mariadb-connector-cpp/X`.

We change `CMAKE_SOURCE_DIR` to `CMAKE_CURRENT_SOURCE_DIR` which expands
to the directory currently being processed by CMake.