CONCPP-120, CONCPP-122 Fix of synchronization and connectTimeout in pool
Added method to blocking_queue to get its lock, and made it used for all
loops thru it to ensure iterators do not get invalidated.
Added timer to connection polling function since timeout give to
wait_for is not guaranteed to be correctly taken(the thread may
spuriously wakeup). Also, there is chance that thread not yet waiting
for condition_variable gets the lock and the connection.
The testcase will be pushed separately.