CONCPP-121 The fix and the testcase
It was possible, that connection returned to the pool as the result of
calling of the close() method, is taken by other
thread(connection object) and before close() metod is finished its work.
So, it could mark underlying Protocol object as closed when it is
actully belonged to othe connection.
The other, but probably less probable possibility, is that since closed
(and returned to the pool) connection held Protocol object, it could
become look as "clean" and usable after other connection gets it from
the pool, and that could lead to the same consequences.
Commit contains new test suite pool2, that runs concurrent pool use,
that reliably repeat this issue in particular.
Also, logging has been added. But it will be brushed up in the separate
commit