CONCPP-123 Adding logging functionality
Connection option "log" can be used to set the logging level and turning
logging on - any value > 0 will turn it on.
"logname" can be used for setting log filename. Default name is
mariadbccpp.log, and it's written to TEMP or USERPROFILE or c: on
Windows, or HOME or /tmp on other systems.
Levels from 1 to 5 are: error, warning, info, debug, trace

Also contains correction of thests in pool2 and small correction of the
pool itself - number of pending requests were not decremented.