CONCPP-129 Property value initialized with NULL was treated as true
Connection properties is map of SQLStrings. If boolean property was initialized
with false, that would treated by compliler as nullptr, and string
initialized with nullptr is empty string, and empty string is treated
for boolean property as true. Which is obviously confusing.

The fix introduces 3rd state for the SQLString - NULL, and makes such
string as a value option to be treated as false. The NULL string will
still have empty string value in different string operations, though.
Setting or changing its value removes the nulliness.
This "Nulliness" is still intended for the driver's internal use mainly,
and designed not to change anything for applications. Comparison of such
object with nullptr returns true, though.